Updated SampleWeb to meet new standards
