+++ src/Sushi.MediaKiwi.Vue/package.json
@@ -3,8 +3,8 @@
   "version": "0.1.3",
   "description": "Vue 3 Library for the MediaKiwi Portal Builder",
   "scripts": {
-    "build": "vue-tsc && vite build",
     "dev": "vue-tsc && vite --config ./vite.config.sample.ts",
+    "api": "cd ../Sushi.MediaKiwi.SampleAPI && dotnet build && dotnet watch run",
     "vi": "vitest",
     "vi:coverage": "vitest --coverage",
     "vi:ui": "vitest --ui",
@@ -14,7 +14,8 @@
     "cy:run:component": "cypress run --component --env coverage=false",
     "cy:coverage:component": "cypress run --component",
     "build:sample": "vue-tsc && vite build --config ./vite.config.sample.ts",
-    "preview:sample": "npm run build:sample && vite preview --config ./vite.config.sample.ts"
+    "preview:sample": "npm run build:sample && vite preview --config ./vite.config.sample.ts",
+    "build": "vue-tsc && vite build"
   },
   "type": "module",
   "files": [
