+++ src/Sushi.MediaKiwi.Vue/sample/models/SampleData.ts
@@ -1,7 +1,7 @@
-export interface ISampleData {
+export type SampleData = {
   id: number;
   name: string;
   countryCode: string;
   countryName: string;
   date: Date;
-}
+};
