--- src/Sushi.MediaKiwi.Vue/sample/views/SampleData.vue
@@ -1,9 +1,0 @@
-<script setup lang="ts">
-  import SampleTable3 from "../components/SampleTable3.vue";
-</script>
-
-<template>
-  <v-card>
-    <SampleTable3></SampleTable3>
-  </v-card>
-</template>
